requirements.txt
streamlit>=1.28.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
PyYAML>=6.0.0
numpy>=1.24.0
requests>=2.31.0
plotly>=5.18.0
python-dotenv>=1.0.0
yfinance>=0.2.0    # S&P 500 (^GSPC); DXY в приложении — FRED DTWEXBGS (см. plan.md §8.5)
jupyter>=1.0.0

# Selenium — fallback для парсинга LookIntoBitcoin (динамические страницы)
selenium>=4.0.0

# Тестирование
pytest>=7.0.0

# P3: реальное исполнение (включите только с BITTREND_LIVE_TRADING + ACK в .env)
ccxt>=4.0.0
